refactor: models are grouped and public class are easy pathed with a good __init__ file
