+++ backend/app/agents/portfolio_agent.py
@@ -14,9 +14,7 @@ from langgraph.graph import END, StateGraph
 from ..models.assets import Asset
 from ..models.portfolio import Portfolio
 from .core.vector_store import VectorStore
-from .state.agent import AgentState
-from .state.analysis import AnalysisResult
-from .state.news import NewsItem
+from ..models import AgentState, AnalysisResult, NewsItem
 from .tools import AnalysisTool, ClassificationTool, NewsSearchTool, PortfolioSummarizerTool
 
 logger = logging.getLogger(__name__)
