+++ backend/app/models/analysis.py
@@ -1,8 +1,10 @@
-# backend/app/agent/state/global.py
+# backend/app/models/analysis.py
 
 from datetime import datetime
 
 from pydantic import BaseModel
+
+from .assets import Asset
 
 
 class NewsItem(BaseModel):
@@ -15,3 +17,13 @@ class NewsItem(BaseModel):
     impact: str | None = None
     relevance_score: float | None = None
     asset_related: str | None = None
+
+
+class AnalysisResult(BaseModel):
+    asset_key: str
+    asset: Asset
+    news_items: list[NewsItem]
+    sentiment_summary: str
+    risk_assessment: str
+    recommendations: list[str]
+    confidence_score: float
